requirements.txt
dash==2.6.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
Flask-Compress==1.13
gunicorn==20.1.0
xarray~=2022.11.0
dask~=2022.10.2
mysql-connector-python~=8.0.31
plotly~=5.10.0
pandas~=1.5.1
numpy~=1.23.4
pip~=22.0.2
colorlog~=4.8.0
wheel~=0.38.2
Brotli~=1.0.9
cryptography~=36.0.2
keyring~=23.5.0
setuptools~=65.5.1
future~=0.18.2
docutils~=0.19
requests~=2.28.1
cdsapi~=0.5.1